+++ tests/tests.R
@@ -276,11 +276,15 @@ connection && Sys.getenv("NOT_CRAN") == "true" && {
   m <- mirai("Seattle", .timeout = 1000)
   if (!is_error_value(m[])) test_equal(m[], "Seattle")
   test_class("errorValue", mirai(q(), .timeout = 1000)[])
-  test_true(daemons(n = 1L, url = local_url(), dispatcher = TRUE))
-  task <- mirai(substitute())
-  url <- nextget("url")
-  test_equal(daemon(url = url, maxtasks = 1L, cleanup = 0L, dispatcher = TRUE), 3L)
-  test_false(daemons(n = 0L))
+  test_true(daemons(sync = TRUE, .compute = "seq"))
+  with_daemons("seq", {
+    test_error(launch_local(), "daemons cannot be launched")
+    test_error(launch_remote(), "synchronous compute profiles")
+    task <- mirai(substitute())
+  })
+  test_false(daemons(0, .compute = "seq"))
+  res <- task[]
+  test_true(missing(res))
   test_type("integer", .Random.seed)
 }
 # TLS tests